+++ crates/objc2/src/__macro_helpers.rs
@@ -642,9 +642,9 @@ mod tests {
 
     #[cfg(feature = "objc2-proc-macros")]
     use crate::__hash_idents;
-    use crate::foundation::{NSDictionary, NSString, NSValue, NSZone};
+    use crate::foundation::{NSDictionary, NSString, NSValue};
     use crate::rc::{Owned, RcTestObject, Shared, ThreadTestData};
-    use crate::runtime::{NSObject, Object};
+    use crate::runtime::{NSObject, NSZone, Object};
     use crate::{class, msg_send_id, ns_string, ClassType};
 
     #[test]